requirements.txt
Django==4.0
python-dotenv==0.20.0
Pillow==9.3.0
djangorestframework==3.12.4
pytz==2021.1
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
django-filter==21.1
webcolors==1.12
